+++ civis/utils/__init__.py
@@ -1,3 +1,4 @@
-from ._jobs import run_job
-
-__all__ = ["run_job"]
+# TODO figure out how to break circular iomport here
+# from ._jobs import run_job
+#
+# __all__ = ["run_job"]